feat(server): permission integration
TODO: refactor root module, get rid of signatures? in root. Move root module to client.

TODO: add ability to generate custom tokens for users. For example, to give applications only access to retrieve transactions. Or for trading bots, to give access only to the balance and the ability to create stock orders. And so on.